Add build script function to ScriptTab
- Rework ScriptTab layout with Build Script button, allow user to build script there, beside the shortcut in Scripting menu.
- Rename editor.dialog package to editor.dialogs.
