+++ src/main/java/editor/SceneTree.java
@@ -3,9 +3,9 @@ package editor;
 import TheCellBeyond.GameObject;
 import TheCellBeyond.GameObject2D;
 import TheCellBeyond.internal.LogicServer;
-import editor.dialog.AddObjectDialog;
-import editor.dialog.ChooseObjectTypeDialog;
-import editor.dialog.SavePrefabDialog;
+import editor.dialogs.AddObjectDialog;
+import editor.dialogs.ChooseObjectTypeDialog;
+import editor.dialogs.SavePrefabDialog;
 import editor.payload.GameObjectDragDropPayload;
 import imgui.ImGui;
 import imgui.ImVec2;
