+++ src/main/java/editor/StartupWindow.java
@@ -4,9 +4,9 @@ import TheCellBeyond.IconLoader;
 import TheCellBeyond.KeyListener;
 import TheCellBeyond.MouseListener;
 import TheCellBeyond.Window;
-import editor.dialog.NewProjectDialog;
-import editor.dialog.OpenProjectDialog;
-import editor.dialog.RemoveMissingProjectDialog;
+import editor.dialogs.NewProjectDialog;
+import editor.dialogs.OpenProjectDialog;
+import editor.dialogs.RemoveMissingProjectDialog;
 import editor.preference.RecentProject;
 import editor.preference.UserPreference;
 import imgui.ImGui;
