+++ src/main/java/editor/dialogs/ChooseObjectTypeDialog.java
@@ -1,4 +1,4 @@
-package editor.dialog;
+package editor.dialogs;
 
 import TheCellBeyond.GameObject;
 import TheCellBeyond.internal.LogicServer;
