+++ src/main/java/editor/dialogs/EditEditorPreferencesDialog.java
@@ -1,4 +1,4 @@
-package editor.dialog;
+package editor.dialogs;
 
 import editor.EditorColors;
 import editor.EditorIcons;
