+++ src/main/java/editor/dialogs/InputMapTab.java
@@ -1,4 +1,4 @@
-package editor.dialog;
+package editor.dialogs;
 
 import TheCellBeyond.*;
 import editor.EditorIcons;
@@ -31,7 +31,7 @@ final class InputMapTab {
     private static boolean showListenForInputDialog = false;
     private static KeyComboCallback pendingCallback = null;
 
-    static boolean isShowListeningDialog() {
+    static boolean openedListenForInput() {
         return showListenForInputDialog;
     }
