+++ src/main/java/editor/dialogs/ListenForInputDialog.java
@@ -1,4 +1,4 @@
-package editor.dialog;
+package editor.dialogs;
 
 import TheCellBeyond.*;
 import editor.EditorLayer;
