+++ src/main/java/editor/dialogs/NewProjectDialog.java
@@ -1,4 +1,4 @@
-package editor.dialog;
+package editor.dialogs;
 
 import editor.EditorColors;
 import editor.preference.RecentProject;
