+++ src/main/java/editor/dialogs/OpenProjectDialog.java
@@ -1,27 +1,34 @@
-package editor.dialog;
+package editor.dialogs;
 
 import org.lwjgl.PointerBuffer;
 import org.lwjgl.system.MemoryStack;
-import org.lwjgl.util.nfd.NFDPickFolderArgs;
+import org.lwjgl.util.nfd.NFDFilterItem;
+import org.lwjgl.util.nfd.NFDOpenDialogArgs;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import static org.lwjgl.system.MemoryUtil.memUTF8;
 import static org.lwjgl.util.nfd.NativeFileDialog.*;
 
-public final class OpenDirectoryDialog extends NativeDialog {
-    public static Path openDialog() {
+public final class OpenProjectDialog extends NativeDialog {
+    public static Path openProjectDialog() {
         setPlatform();
         try (MemoryStack stack = MemoryStack.stackPush()) {
-            PointerBuffer outPath = stack.mallocPointer(1);
-            int result = NFD_PickFolder_With(outPath, NFDPickFolderArgs.calloc(stack)
+            NFDFilterItem.Buffer filter = NFDFilterItem.malloc(1);
+            filter.get(0)
+                    .name(stack.UTF8("_project"))
+                    .spec(stack.UTF8("yml,yaml"));
+            PointerBuffer pointerBuffer = stack.mallocPointer(1);
+            int result = NFD_OpenDialog_With(pointerBuffer, NFDOpenDialogArgs.calloc(stack)
+                    .filterList(filter)
                     .parentWindow(it -> it
                             .type(handleType)
                             .handle(windowHandle)
                     )
             );
-            return checkResult(result, outPath);
+            return checkResult(result, pointerBuffer);
         } catch (Exception e) {
-            System.err.println("Error while opening directory dialog: " + e.getMessage());
+            System.err.println("Error while opening file dialog: " + e.getMessage());
             return null;
         }
     }
@@ -32,11 +39,11 @@ public final class OpenDirectoryDialog extends NativeDialog {
             case NFD_OKAY -> {
                 long pathPtr = pp.get(0);
                 String selectedPath = memUTF8(pathPtr);
-                if (validDirectory(selectedPath)){
+                if (isFileValid(selectedPath)){
                     validPath = Paths.get(selectedPath);
                     NFD_FreePath(pathPtr);
                 } else {
-                    System.out.println("Selected directory does not exist");
+                    System.out.println("Selected file format not supported.");
                     NFD_FreePath(pathPtr);
                 }
             }
@@ -46,9 +53,8 @@ public final class OpenDirectoryDialog extends NativeDialog {
         return validPath;
     }
 
-    private static boolean validDirectory(String path) {
-        if (path == null || path.isEmpty()) return false;
-        Path check = Path.of(path);
-        return check.toFile().isDirectory();
+    private static boolean isFileValid(String path) {
+        if (path == null || path.isEmpty() || !Files.exists(Paths.get(path))) return false;
+        return path.toLowerCase().endsWith(".yml") || path.toLowerCase().endsWith(".yaml");
     }
 }
