+++ src/main/java/editor/dialogs/SaveSceneAsDialog.java
@@ -1,6 +1,5 @@
-package editor.dialog;
+package editor.dialogs;
 
-import TheCellBeyond.internal.LogicServer;
 import imgui.ImGui;
 import imgui.ImVec2;
 import imgui.flag.ImGuiCol;
@@ -9,8 +8,8 @@ import imgui.flag.ImGuiWindowFlags;
 import imgui.type.ImString;
 import scene.SceneManager;
 
-public final class RenameSceneDialog {
-    private static final String PopupID = "Rename Scene##TCB_Rename_Scene_Dialog";
+public final class SaveSceneAsDialog {
+    private static final String PopupID = "Save scene as...##TCB_Save_Scene_As_Dialog";
     private static final ImVec2 DialogSize = new ImVec2(400.0f, 160.0f);
     private static final float ButtonReserve = ImGui.getFrameHeightWithSpacing();
     private static final float ButtonWidth = 100.0f;
@@ -19,19 +18,19 @@ public final class RenameSceneDialog {
     private static boolean nameTaken = false;
     private static final ImString sceneName = new ImString(128);
     private static String errorMessage = "";
-
-    private RenameSceneDialog() {}
+    private static Runnable onSaveCallback = null;
 
     public static void show() {
-        String current = LogicServer.currentSceneName();
-        if (current == null) return;
-        showDialog = true;
-        sceneName.set(current);
-        nameTaken = false;
-        errorMessage = "";
+        show(null);
     }
 
-    public static void imgui() {
+    public static void show(Runnable callback) {
+        resetDialogData();
+        showDialog = true;
+        onSaveCallback = callback;
+    }
+
+    public static void imgui()  {
         if (!showDialog) return;
         ImGui.openPopup(PopupID);
         ImVec2 centre = ImGui.getMainViewport().getCenter();
@@ -40,30 +39,32 @@ public final class RenameSceneDialog {
         ImGui.setNextWindowSize(DialogSize);
         if (ImGui.beginPopupModal(PopupID, ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoScrollbar)) {
             ImGui.spacing();
-            ImGui.text("New name:");
+            ImGui.text("Scene's name:");
             ImGui.spacing();
             ImGui.pushItemWidth(ImGui.getContentRegionAvailX());
-            if (ImGui.inputTextWithHint("##RenameSceneInput", "Enter a new name...", sceneName)) checkName();
+            if (ImGui.inputTextWithHint("##SSAD_Scene_Name_Input", "Enter a name for the scene...", sceneName)) checkNewSceneName();
             ImGui.popItemWidth();
             if (nameTaken || !errorMessage.isEmpty()) {
                 ImGui.pushStyleColor(ImGuiCol.Text, 1.0f, 0.2f, 0.2f, 1.0f);
                 ImGui.textWrapped(errorMessage);
                 ImGui.popStyleColor(1);
-            } else ImGui.text("    ");
+            } else {
+                ImGui.text("    ");
+            }
             ImGui.setCursorPosY(ImGui.getWindowHeight() - ButtonReserve - ButtonHeight / 2.0f);
             float buttonPivotX = ButtonWidth * 0.5f;
             float startX = ImGui.getCursorStartPosX();
             float availX = ImGui.getContentRegionAvailX();
-            float createX = startX + availX * 0.25f - buttonPivotX;
+            float saveX = startX + availX * 0.25f - buttonPivotX;
             float cancelX = startX + availX * 0.75f - buttonPivotX;
-            boolean canRename = !sceneName.isEmpty() && !nameTaken && errorMessage.isEmpty();
-            ImGui.setCursorPosX(createX);
-            if (!canRename) ImGui.beginDisabled();
-            if (ImGui.button("Rename##RSD_Confirm_Rename_Scene_Button", ButtonWidth, ButtonHeight)) rename();
-            if (!canRename) ImGui.endDisabled();
+            boolean canSave = !sceneName.isEmpty() && !nameTaken && errorMessage.isEmpty();
+            ImGui.setCursorPosX(saveX);
+            if (!canSave) ImGui.beginDisabled();
+            if (ImGui.button("Save##SSAD_Save_As_Button", ButtonWidth, ButtonHeight)) saveScene();
+            if (!canSave) ImGui.endDisabled();
             ImGui.sameLine();
             ImGui.setCursorPosX(cancelX);
-            if (ImGui.button("Cancel##RSD_Cancel_Rename_Scene_Button", ButtonWidth, ButtonHeight)) {
+            if (ImGui.button("Cancel##SSAD_Cancel_Save_Button", ButtonWidth, ButtonHeight)) {
                 showDialog = false;
                 ImGui.closeCurrentPopup();
             }
@@ -75,51 +76,46 @@ public final class RenameSceneDialog {
         }
     }
 
-    private static void checkName() {
+    private static void resetDialogData() {
+        sceneName.clear();
+        nameTaken = false;
+        onSaveCallback = null;
+        errorMessage = "";
+    }
+
+    private static void checkNewSceneName() {
         String name = sceneName.get().trim();
         if (!SceneManager.validSceneName(name)) {
             nameTaken = false;
-            errorMessage = "Name cannot be blank";
-            return;
-        }
-        String currentName = LogicServer.currentSceneName();
-        if (name.equals(currentName)) {
-            nameTaken = false;
-            errorMessage = "";
+            errorMessage = "Name cannot be empty";
             return;
         }
         if (SceneManager.sceneNameTaken(name)) {
             nameTaken = true;
-            errorMessage = "Scene '" + name + "' already exist";
+            errorMessage = "Scene '" + name + "' already existed";
             return;
         }
         nameTaken = false;
         errorMessage = "";
     }
 
-    private static void rename() {
+    private static void saveScene() {
         String name = sceneName.get().trim();
-        String currentName = LogicServer.currentSceneName();
         if (!SceneManager.validSceneName(name) || nameTaken) {
             errorMessage = "Entered name is empty or already taken";
             return;
         }
-        if (name.equals(currentName)) {
-            showDialog = false;
-            ImGui.closeCurrentPopup();
+        if (!SceneManager.saveSceneAs(name)) {
+            errorMessage = "Failed to save new scene '" + name + "'";
             return;
         }
-        if (!SceneManager.renameScene(currentName, name)) {
-            errorMessage = "Failed to rename scene";
-            return;
+        if (onSaveCallback != null) {
+            Runnable callback = onSaveCallback;
+            onSaveCallback = null;
+            callback.run();
         }
         showDialog = false;
         ImGui.closeCurrentPopup();
     }
 
-    private static void resetDialogData() {
-        nameTaken = false;
-        sceneName.clear();
-        errorMessage = "";
-    }
 }
