+++ src/main/java/editor/dialogs/ScriptsTab.java
@@ -1,10 +1,12 @@
-package editor.dialog;
+package editor.dialogs;
 
 import editor.EditorColors;
 import editor.EditorIcons;
 import editor.EditorWidget;
 import editor.preference.EditorPreferences;
 import editor.preference.UserPreference;
+import editor.widgets.SelectableTextView;
+import editor.widgets.SelectableTextView.Row;
 import imgui.ImGui;
 import imgui.ImVec2;
 import imgui.flag.*;
@@ -12,11 +14,13 @@ import imgui.type.ImString;
 import project.Project;
 import scripting.ScriptLoader;
 import scripting.ScriptProjectGenerator;
+import scripting.builder.BuildPhase;
+import scripting.builder.BuildStatus;
 import scripting.builder.ScriptBuilder;
 import utility.log.EngineLog;
 
+import java.util.ArrayList;
 import java.util.List;
-
 
 final class ScriptsTab {
     private enum HoveredControl {
@@ -26,23 +30,33 @@ final class ScriptsTab {
         GenerateScriptProject,
     }
 
-    private static final float Padding = 4.0f;
     private static final ImString newScanDirName = new ImString(256);
     private static final ImVec2 sizeCache = new ImVec2();
-    private static final float IconSize = 28.0f;
+
     private static final String BuildScriptLabel = "Build Script##ST_Build_Script_Control_Button";
     private static final String UpdateScriptProjectLabel = "Update Script Project##ST_Update_Script_Project_Control_Button";
     private static final String GenerateScriptProjectButton = "Generate Script Project##ST_Generate_Script_Project_Control_Button";
+    private static final float Padding = 4.0f;
     private static final float ButtonWidth = Math.max(ImGui.calcTextSizeX("Update Script Project"), ImGui.calcTextSizeX("Generate Script Project")) + Padding + ImGui.getStyle().getFramePaddingX();
     private static final float ButtonHeight = 30.0f;
     private static final float ErrorReserveRegionHeight = ImGui.getTextLineHeightWithSpacing() * 5.0f;
+    private static final float IconSize = 28.0f;
+
     private static float scanDirLayoutInnerHeight = IconSize;
     private static float mainScriptControlLayoutInnerHeight = ButtonHeight * 3.0f;
     private static HoveredControl hoveredControl = HoveredControl.None;
 
+    private static String message = "";
+    private static final SelectableTextView ErrorView = new SelectableTextView();
+    private static final List<Row> ErrorRows = new ArrayList<>();
+    private static BuildStatus lastStatus;
+    private static boolean hasError = false;
+
     static void clear() {
         hoveredControl = HoveredControl.None;
         newScanDirName.clear();
+        message = "";
+        hasError = false;
     }
 
     static void imgui() {
@@ -81,8 +95,10 @@ final class ScriptsTab {
         if (!canAdd) ImGui.endDisabled();
         ImGui.endTable();
         ImGui.separator();
+        refreshStatus();
         float remainHeight = ImGui.getContentRegionAvailY();
-        if (ImGui.beginChild("##ST_ScanDir_List_Region", 0.0f, remainHeight - sizeCache.y - ErrorReserveRegionHeight - Padding)) renderScanDirList();
+        float buildStatusReserve = hasError ? ErrorReserveRegionHeight : ImGui.getTextLineHeightWithSpacing();
+        if (ImGui.beginChild("##ST_ScanDir_List_Region", 0.0f, remainHeight - sizeCache.y - buildStatusReserve - Padding)) renderScanDirList();
         ImGui.endChild();
         ImGui.beginGroup();
         ImGui.separator();
@@ -93,6 +109,12 @@ final class ScriptsTab {
         renderScriptControl();
         ImGui.endGroup();
         ImGui.getItemRectSize(sizeCache);
+        if (!hasError) {
+            EditorWidget.textCenterAlign(message);
+            return;
+        }
+        if (ImGui.beginChild("##ST_Build_Error_View_Region", ImGuiChildFlags.Borders)) ErrorView.render("##ST_Build_Error_View", ErrorRows);
+        ImGui.endChild();
     }
 
     private static void renderClearButton(Runnable onClear, float height) {
@@ -122,10 +144,7 @@ final class ScriptsTab {
     }
 
     private static void renderScriptControl() {
-        if (!ImGui.beginTable("ST_Main_Script_Control_Table_Layout", 2, ImGuiTableFlags.SizingFixedFit, ImGui.getContentRegionAvailX(), mainScriptControlLayoutInnerHeight)) {
-            ImGui.endGroup();
-            return;
-        }
+        if (!ImGui.beginTable("ST_Main_Script_Control_Table_Layout", 2, ImGuiTableFlags.SizingFixedFit, ImGui.getContentRegionAvailX(), mainScriptControlLayoutInnerHeight)) return;
         ImGui.tableSetupColumn("ST_Main_Script_Control_Buton_Column", ImGuiTableColumnFlags.WidthFixed);
         ImGui.tableSetupColumn("ST_Main_Script_Control_Description_Column", ImGuiTableColumnFlags.WidthStretch);
         ImGui.tableNextColumn();
@@ -149,7 +168,6 @@ final class ScriptsTab {
         ImGui.endChild();
         ImGui.endTable();
         ImGui.separator();
-
     }
 
     private static void renderDescription() {
@@ -158,7 +176,7 @@ final class ScriptsTab {
             case BuildScript -> {
                 ImGui.textWrapped("Click \"Build Script\" to compile script classes into JAR and load them into the engine.");
                 ImGui.textWrapped("All .tcbs script file will be translated to .java source first.");
-                ImGui.textWrapped("This will also reload the current scene if that option is enabled in Editor Preference");
+                ImGui.textWrapped("This will also reload the current scene if that option is enabled in Editor Preferences.");
             }
             case UpdateScriptProject -> {
                 ImGui.textWrapped("Click \"Update Script Project\" to update the API version to the latest version supported by the engine.");
@@ -186,7 +204,48 @@ final class ScriptsTab {
     }
 
     private static void requestBuildScript() {
+        if (ScriptBuilder.inProgress()) return;
+        message = "Starting script project build task...";
         EditorPreferences preferences = UserPreference.preferences();
         ScriptBuilder.build(UserPreference.selectedJDKHome(), preferences.cleanBuildScripts(), preferences.overrideGradleJVM(), preferences.reloadOnFinishBuildScripts());
+    }
+
+    private static void refreshStatus() {
+        BuildStatus status = ScriptBuilder.status();
+        if (status == lastStatus) return;
+        lastStatus = status;
+        hasError = status.phase() == BuildPhase.TranspilerFailed || status.phase() == BuildPhase.BuildFailed;
+        if (hasError) {
+            rebuildErrorRows(status);
+            return;
+        }
+        message = switch (status.phase()) {
+            case Transpiling -> "Translating .tcbs script files...";
+            case Building -> String.format("Translated %d .tcbs files. Building script project...", status.scriptCount());
+            case Succeeded -> String.format("Build script project finished in %s (Exit code %d)", formatDuration(status.buildMS()), status.exitCode());
+            default -> message;
+        };
+    }
+
+    private static void rebuildErrorRows(BuildStatus status) {
+        ErrorRows.clear();
+        ErrorView.clearSelection();
+        if (status.errors().isEmpty()) {
+            ErrorRows.add(new Row(String.format("Script build failed (exit code %d), check the Gradle output in the console", status.exitCode()), EditorColors.ErrorLogColor));
+            return;
+        }
+        status.errors().forEach(e -> ErrorRows.add(new Row(e, EditorColors.ErrorLogColor)));
+    }
+
+    private static String formatDuration(long durationMs) {
+        if (durationMs < 1000L) return durationMs + "ms";
+        long hours = durationMs / 3_600_000L;
+        long minutes = (durationMs / 60_000L) % 60L;
+        long seconds = (durationMs / 1000L) % 60L;
+        StringBuilder elapsed = new StringBuilder();
+        if (hours > 0) elapsed.append(hours).append("h ");
+        if (minutes > 0) elapsed.append(minutes).append("m ");
+        if (seconds > 0) elapsed.append(seconds).append("s");
+        return elapsed.toString().trim();
     }
 }
